secureCasualConversationScraper
Scrapes the top of all time posts for a given subreddit. (hardcoded)

The api calls and file calls are correct but there is an issue passing the strings containing the key and passwords from a file. (possibly formatting)
This can be solved by hardcoding however it would not be suitable for public upload